organizando tudo desde o comeco

+++ perfil/includes/documentos_fiscais_incentivador_pf.php
@@ -4,56 +4,66 @@ $idPf = $_SESSION['idUser'];
 $tipoPessoa = '4';
 
 if (isset($_POST['iniciar_incentivo'])) {
-    $sqlEtapa = "INSERT INTO etapas_incentivo (tipoPessoa, idIncentivador, etapa) VALUES ($tipoPessoa, $idPf, 1)";
+    $sqlEtapa = "INSERT INTO incentivador_projeto (tipoPessoa, idPessoa, etapa) VALUES ($tipoPessoa, $idPf, 1)";
     if (mysqli_query($con, $sqlEtapa)) {
+        $_SESSION['idIncentivadorProjeto'] = recuperaUltimo("incentivador_projeto");
+        $idIncentivadorProjeto = $_SESSION['idIncentivadorProjeto'];
+        echo $_SESSION['idIncentivadorProjeto'];
         $mensagemInicial = "<font color='#01DF3A'><strong>Você iniciou o processo de incentivar um projeto.<br>Por favor, siga as etapas seguintes preenchendo corretamente todas as informações solicitadas.</strong></font>";
     } else {
         echo $sqlEtapa;
     }
+} else {
+    $idIncentivadorProjeto = $_SESSION['idIncentivadorProjeto'];
+    echo $_SESSION['idIncentivadorProjeto'];
 }
 
 if (isset($_POST["enviar"])) {
     $sql_arquivos = "SELECT * FROM lista_documento WHERE idTipoUpload = '3' AND idListaDocumento IN (39, 40, 41, 42, 43, 54)";
     $query_arquivos = mysqli_query($con, $sql_arquivos);
     while ($arq = mysqli_fetch_array($query_arquivos)) {
-        $y = $arq['idListaDocumento'];
-        $x = $arq['sigla'];
-        $nome_arquivo = isset($_FILES['arquivo']['name'][$x]) ? $_FILES['arquivo']['name'][$x] : null;
-        $f_size = isset($_FILES['arquivo']['size'][$x]) ? $_FILES['arquivo']['size'][$x] : null;
+        if (!verificaArquivosExistentesIncentivador($idPf, $arq['idListaDocumento'], 3) ) {
+            $y = $arq['idListaDocumento'];
+            $x = $arq['sigla'];
+            $nome_arquivo = isset($_FILES['arquivo']['name'][$x]) ? $_FILES['arquivo']['name'][$x] : null;
+            $f_size = isset($_FILES['arquivo']['size'][$x]) ? $_FILES['arquivo']['size'][$x] : null;
 
-        //Extensões permitidas
-        $ext = array("PDF", "pdf");
+            //Extensões permitidas
+            $ext = array("PDF", "pdf");
 
-        if ($f_size > 6242880) // 6MB em bytes
-        {
-            $mensagem = "<font color='#FF0000'><strong>Erro! Tamanho de arquivo excedido! Tamanho máximo permitido: 06 MB.</strong></font>";
-        } else {
-            if ($nome_arquivo != "") {
-                $nome_temporario = $_FILES['arquivo']['tmp_name'][$x];
-                $new_name = date("YmdHis") . "_" . semAcento($nome_arquivo); //Definindo um novo nome para o arquivo
-                $hoje = date("Y-m-d H:i:s");
-                $dir = '../uploadsdocs/'; //Diretório para uploads
-                $allowedExts = array(".pdf", ".PDF"); //Extensões permitidas
-                $ext = strtolower(substr($nome_arquivo, -4));
+            if ($f_size > 6242880) // 6MB em bytes
+            {
+                $mensagem = "<font color='#FF0000'><strong>Erro! Tamanho de arquivo excedido! Tamanho máximo permitido: 06 MB.</strong></font>";
+            } else {
+                if ($nome_arquivo != "") {
+                    $nome_temporario = $_FILES['arquivo']['tmp_name'][$x];
+                    $new_name = date("YmdHis") . "_" . semAcento($nome_arquivo); //Definindo um novo nome para o arquivo
+                    $hoje = date("Y-m-d H:i:s");
+                    $dir = '../uploadsdocs/'; //Diretório para uploads
+                    $allowedExts = array(".pdf", ".PDF"); //Extensões permitidas
+                    $ext = strtolower(substr($nome_arquivo, -4));
 
-                if (in_array($ext, $allowedExts)) //Pergunta se a extensão do arquivo, está presente no array das extensões permitidas
-                {
-                    if (move_uploaded_file($nome_temporario, $dir . $new_name)) {
-                        $sql_insere_arquivo = "INSERT INTO `upload_arquivo` (`idTipo`, `idPessoa`, `idListaDocumento`, `arquivo`, `dataEnvio`, `publicado`) VALUES ('$tipoPessoa', '$idPf', '$y', '$new_name', '$hoje', '1'); ";
-                        $query = mysqli_query($con, $sql_insere_arquivo);
-                        if ($query) {
-                            $mensagem = "<font color='#01DF3A'><strong>Arquivo(s) recebido(s) com sucesso!</strong></font>";
-                            gravarLog($sql_insere_arquivo);
+                    if (in_array($ext, $allowedExts)) //Pergunta se a extensão do arquivo, está presente no array das extensões permitidas
+                    {
+                        if (move_uploaded_file($nome_temporario, $dir . $new_name)) {
+                            $sql_insere_arquivo = "INSERT INTO `upload_arquivo` (`idTipo`, `idPessoa`, `idListaDocumento`, `arquivo`, `dataEnvio`, `publicado`) VALUES ('3', '$idPf', '$y', '$new_name', '$hoje', '1'); ";
+                            $query = mysqli_query($con, $sql_insere_arquivo);
+                            if ($query) {
+                                $mensagem = "<font color='#01DF3A'><strong>Arquivo(s) recebido(s) com sucesso!</strong></font>";
+                                gravarLog($sql_insere_arquivo);
+                            } else {
+                                $mensagem = "<font color='#FF0000'><strong>Erro ao gravar no banco.</strong></font>";
+                            }
                         } else {
-                            $mensagem = "<font color='#FF0000'><strong>Erro ao gravar no banco.</strong></font>";
+                            $mensagem = "<font color='#FF0000'><strong>Erro no upload! Tente novamente.</strong></font>";
                         }
                     } else {
-                        $mensagem = "<font color='#FF0000'><strong>Erro no upload! Tente novamente.</strong></font>";
+                        $mensagem = "<font color='#FF0000'><strong>Erro no upload! Anexar documentos somente no formato PDF.</strong></font>";
                     }
-                } else {
-                    $mensagem = "<font color='#FF0000'><strong>Erro no upload! Anexar documentos somente no formato PDF.</strong></font>";
                 }
             }
+        } else {
+            echo "<script> swal('Você já anexou o " . $arq['documento'] . "', '', 'warning') </script>";
         }
     }
 }
@@ -70,6 +80,21 @@ if (isset($_POST['apagar'])) {
 }
 
 $pf = recuperaDados("incentivador_pessoa_fisica", "idPf", $idPf);
+
+$docs = ['39, 40, 41, 42, 43, 54'];
+
+foreach ($docs as $doc) {
+    if (verificaArquivosExistentesIncentivador($idPf, $doc, 3)) {
+        $arqAnexados = 'block';
+        break;
+    } else {
+        $arqAnexados = 'none';
+    }
+}
+
+
+
+
 
 ?>
 <section id="list_items" class="home-section bg-white">
@@ -108,17 +133,18 @@ $pf = recuperaDados("incentivador_pessoa_fisica", "idPf", $idPf);
                 {
                 ?>
                 <!-- Exibir arquivos -->
-                <div class="form-group">
+                <div class="form-group" style="display: <?=$arqAnexados?>">
                     <div class="col-md-12">
                         <div class="table-responsive list_info"><h6>Arquivo(s) Anexado(s)</h6>
                             <?php
 
-                            $arqsEnviados = listaArquivosPessoa($idPf, $tipoPessoa, "includes/documentos_fiscais_incentivador_pf", "39, 40, 41, 42, 43, 54");
+                            $arqsEnviados = listaArquivosPessoa($idPf, 3, "includes/documentos_fiscais_incentivador_pf", "39, 40, 41, 42, 43, 54");
                             if ($arqsEnviados == 6) {
                                 echo "
                                       <form method='POST' action='?perfil=includes/incentivadorPF_etapa3_visualiza_docs' enctype='multipart/form-data'>
                                       <input type='hidden' name='idPf' value='$idPf'>                                   
-                                          <input type='submit' name='enviarSMC' class='btn btn-theme btn-lg btn-block'
+                                      <input type='hidden' name='idIncentivadorProjeto' value='$idIncentivadorProjeto'>    
+                                      <input type='submit' name='enviarSMC' class='btn btn-theme btn-lg btn-block'
                                                value='Enviar à SMC'>                                     
                                         
                                       </form>  ";
@@ -139,27 +165,29 @@ $pf = recuperaDados("incentivador_pessoa_fisica", "idPf", $idPf);
                                 $query_arquivos = mysqli_query($con, $sql_arquivos);
                                 while ($arq = mysqli_fetch_array($query_arquivos)) {
                                     $doc = $arq['documento'];
-                                    $query = "SELECT idListaDocumento FROM lista_documento WHERE documento='$doc' AND publicado='1' AND idTipoUpload='3'";
-                                    $envio = $con->query($query);
-                                    $row = $envio->fetch_array(MYSQLI_ASSOC);
+                                    $idListaDoc = $arq['idListaDocumento'];
 
-                                    if (!verificaArquivosExistentesIncentivador($idPf, $row['idListaDocumento'])) {
+                                    if (!verificaArquivosExistentesIncentivador($idPf, $idListaDoc, 3)) {
                                         $documento = (object)
                                         [
                                             'nomeDocumento' => $arq['documento'],
                                             'sigla' => $arq['sigla']
                                         ];
                                         array_push($documentos, $documento);
+
                                     }
                                 }
 
                                 if ($documentos) {
                                     ?>
                                     <table class='table table-condensed table-striped'>
-                                        <tr class='list_menu'>
-                                            <td>Tipo de Arquivo</td>
-                                            <td></td>
-                                        </tr>
+                                        <thead>
+                                            <tr class='list_menu'>
+                                                <td>Tipo de Arquivo</td>
+                                                <td></td>
+                                            </tr>
+                                        </thead>
+                                        <tbody>
                                         <?php
                                         foreach ($documentos as $documento) {
                                             echo "<tr>";
@@ -168,6 +196,7 @@ $pf = recuperaDados("incentivador_pessoa_fisica", "idPf", $idPf);
                                             echo "<tr>";
                                         }
                                         ?>
+                                        </tbody>
                                     </table>
                                     <input type="hidden" name="idPessoa" value="<?php echo $idPf; ?>"/>
                                     <input type="hidden" name="tipoPessoa" value="<?php echo $tipoPessoa; ?>"/>
